requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
phi
numpy>=1.21.0
pandas>=1.3.0
requests>=2.26.0
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit
phidata
google-generativeai
openai
duckduckgo-search

